fix: fix  data not saving to database

+++ app/src/main/java/com/ezzy/weatherapptest/data/remote/dto/current_weather_hourly/Minutely.kt
@@ -2,9 +2,9 @@ package com.ezzy.weatherapptest.data.remote.dto.current_weather_hourly
 
 data class Minutely(
     val precip: Int,
-    val snow: Int,
+    val snow: Float,
     val temp: Double,
     val timestamp_local: String,
     val timestamp_utc: String,
-    val ts: Int
+    val ts: Float
 )
